# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory Management API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/inventory": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Get all inventory items",
                "operationId": "4c6266820dcc86d9c9d5d6ca66eb4ec8",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/AcceptHeader"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inventory list"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Create an inventory record",
                "operationId": "a8574036500a367c4a9a07c4f9aa9a44",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/AcceptHeader"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "warehouse_id",
                                    "minimum_quantity"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "warehouse_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "minimum_quantity": {
                                        "type": "integer",
                                        "minimum": 0,
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Inventory created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/inventory/{id}": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Get inventory by ID",
                "operationId": "a0043df6870243ed184a03cdefab4e1a",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/AcceptHeader"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inventory details"
                    },
                    "404": {
                        "description": "Inventory not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Update inventory record",
                "operationId": "ecf874bde3fb3ff5ed0e7d65b4fba310",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/AcceptHeader"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "warehouse_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "minimum_quantity": {
                                        "type": "integer",
                                        "minimum": 0,
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inventory updated"
                    },
                    "404": {
                        "description": "Inventory not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Delete inventory record",
                "operationId": "47d1094b944bf5e60333904658030088",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/AcceptHeader"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Inventory deleted"
                    },
                    "404": {
                        "description": "Inventory not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/inventories/global-view": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Get global inventory view",
                "operationId": "c498aa341a0e9c422a222080b30063b8",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/AcceptHeader"
                    },
                    {
                        "name": "warehouse_id",
                        "in": "query",
                        "description": "Filter by warehouse ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "Filter by product ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Global inventory summary"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/inventory-transfer": {
            "post": {
                "tags": [
                    "Inventory Transaction"
                ],
                "summary": "Transfer inventory from one warehouse to another",
                "operationId": "ba76ab9367e8e7466c43c25cd106d8cd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "from_warehouse_id",
                                    "to_warehouse_id",
                                    "quantity",
                                    "date"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 7
                                    },
                                    "from_warehouse_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "to_warehouse_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "minimum": 1,
                                        "example": 50
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-07-18"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inventory transfer completed"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/report/low-stock": {
            "get": {
                "tags": [
                    "Report"
                ],
                "summary": "Get report of products with low stock",
                "operationId": "005d84b5d1deaa441c007b71689c7897",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/AcceptHeader"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of low stock products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "product_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "warehouse_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "quantity": {
                                                "type": "integer",
                                                "example": 100
                                            },
                                            "minimum_quantity": {
                                                "type": "integer",
                                                "example": 102
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-07-14T22:12:50.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-07-16T17:59:58.000000Z"
                                            },
                                            "deleted_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": null,
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "parameters": {
            "AcceptHeader": {
                "name": "Accept",
                "in": "header",
                "description": "Force response format to JSON",
                "required": true,
                "schema": {
                    "type": "string",
                    "default": "application/json"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}